chore: reorganize added files
