+++ src/vs/workbench/contrib/void/electron-main/llmMessage/sendLLMMessage.ts
@@ -6,7 +6,7 @@
 import { SendLLMMessageParams, OnText, OnFinalMessage, OnError } from '../../common/sendLLMMessageTypes.js';
 import { IMetricsService } from '../../common/metricsService.js';
 import { displayInfoOfProviderName } from '../../common/voidSettingsTypes.js';
-import { getProvider } from './providers/index.js';
+import { getProvider } from './modelProvider.js';
 
 
 export const sendLLMMessage = async ({
